+++ packages/icons-react/src/icons/CheckMark.tsx
@@ -1,8 +1,8 @@
-import React from "react";
+import React, { FC } from "react";
 import { IconFactory } from "../IconFactory";
 import { IconProps } from "./types";
 
-export const CheckMark: React.FC<IconProps> = ({ className, variant }) => (
+export const CheckMark: FC<IconProps> = ({ className, variant }) => (
     <IconFactory
         title="CheckMark"
         description="Ikon av en hake"