+++ packages/icons-react/src/icons/Warning.tsx
@@ -1,8 +1,8 @@
-import React from "react";
+import React, { FC } from "react";
 import { IconFactory } from "../IconFactory";
 import { IconProps } from "./types";
 
-export const Warning: React.FC<IconProps> = ({ className, variant }) => (
+export const Warning: FC<IconProps> = ({ className, variant }) => (
     <IconFactory
         viewBox="0 0 24 24"
         title="Warning-icon"